+++ src/Services/BaseService.php
@@ -4,17 +4,22 @@ namespace DreamFactory\Core\Email\Services;
 
 use App;
 use DreamFactory\Core\Contracts\EmailServiceInterface;
+use DreamFactory\Core\Email\Components\Attachment;
 use DreamFactory\Core\Email\Components\EmailUtilities;
 use DreamFactory\Core\Email\Components\Mailer as DfMailer;
+use DreamFactory\Core\Enums\Verbs;
+use DreamFactory\Core\Exceptions\ForbiddenException;
 use DreamFactory\Core\Exceptions\NotFoundException;
 use DreamFactory\Core\Exceptions\BadRequestException;
 use DreamFactory\Core\Exceptions\InternalServerErrorException;
 use DreamFactory\Core\Models\EmailTemplate;
 use DreamFactory\Core\Services\BaseRestService;
+use DreamFactory\Core\Utility\FileUtilities;
 use DreamFactory\Core\Utility\Session;
 use Illuminate\Mail\Message;
 use Swift_Transport as SwiftTransport;
 use Swift_Mailer as SwiftMailer;
+use ServiceManager;
 
 abstract class BaseService extends BaseRestService implements EmailServiceInterface
 {
@@ -55,6 +60,9 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
      */
     abstract protected function setTransport($config);
 
+    /**
+     * @throws \DreamFactory\Core\Exceptions\InternalServerErrorException
+     */
     protected function setMailer()
     {
         if (!$this->transport instanceof SwiftTransport) {
@@ -65,6 +73,9 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
         $this->mailer = new DfMailer(App::make('view'), $swiftMailer, App::make('events'));
     }
 
+    /**
+     * @param $config
+     */
     protected function setParameters($config)
     {
         $this->parameters = (array)array_get($config, 'parameters', []);
@@ -83,6 +94,159 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
     }
 
     /**
+     * Gets uploaded file(s) for attachment.
+     *
+     * @param null|string|array $path
+     *
+     * @return array
+     */
+    protected function getUploadedAttachment($path = null)
+    {
+        $attachment = [];
+        $file = $this->request->getFile('file', $this->request->getFile('attachment', $path));
+
+        if (is_array($file)) {
+            if (isset($file['tmp_name'], $file['name'])) {
+                $attachment[] = new Attachment($file['tmp_name'], array_get($file, 'name'));
+            } else {
+                foreach ($file as $f) {
+                    if (isset($f['tmp_name'], $f['name'])) {
+                        $attachment[] = new Attachment(array_get($f, 'tmp_name'), array_get($f, 'name'));
+                    }
+                }
+            }
+        }
+
+        return $attachment;
+    }
+
+    /**
+     * Gets URL imported file(s) for attachment.
+     *
+     * @param null|string $path
+     *
+     * @return array
+     * @throws \DreamFactory\Core\Exceptions\InternalServerErrorException
+     */
+    protected function getUrlAttachment($path = null)
+    {
+        $attachment = [];
+        $file = $this->request->input('import_url', $this->request->input('attachment', $path));
+
+        if (!empty($file)) {
+            if (!is_array($file)) {
+                $files = explode(',', $file);
+            } else {
+                $files = $file;
+            }
+            try {
+                foreach ($files as $f) {
+                    if (is_string($f)) {
+                        Session::replaceLookups($f);
+                        $fileURL = urldecode($f);
+                        $filePath = FileUtilities::importUrlFileToTemp($fileURL);
+                        $attachment[] = new Attachment($filePath, basename($fileURL));
+                    }
+                }
+            } catch (\Exception $e) {
+                throw new InternalServerErrorException('Failed to import attachment file from url. ' .
+                    $e->getMessage());
+            }
+        }
+
+        return $attachment;
+    }
+
+    /**
+     * Gets file(s) stored in storage service(s) for attachment.
+     *
+     * @param null|string|array $path
+     *
+     * @return array
+     * @throws \DreamFactory\Core\Exceptions\InternalServerErrorException
+     */
+    protected function getServiceAttachment($path = null)
+    {
+        $attachment = [];
+        $file = $this->request->input('import_url', $this->request->input('attachment', $path));
+
+        if (!empty($file) && is_array($file)) {
+            if (isset($file['service'])) {
+                $files = [$file];
+            } else {
+                $files = $file;
+            }
+
+            try {
+                foreach ($files as $f) {
+                    if (is_array($f)) {
+                        $service = array_get($f, 'service');
+                        $path = array_get($f, 'path', array_get($f, 'file_path'));
+                        Session::replaceLookups($service);
+                        Session::replaceLookups($path);
+
+                        if (empty($service) || empty($path)) {
+                            throw new BadRequestException('No service name and file path provided in request.');
+                        }
+
+                        if (Session::checkServicePermission(Verbs::GET, $service, $path, Session::getRequestor(),
+                            false)) {
+                            /** @var \DreamFactory\Core\Contracts\ServiceResponseInterface $result */
+                            $result = ServiceManager::handleRequest(
+                                $service,
+                                Verbs::GET,
+                                $path,
+                                ['include_properties' => true, 'content' => true, 'is_base64' => true]
+                            );
+
+                            if ($result->getStatusCode() !== 200) {
+                                throw new InternalServerErrorException(
+                                    'Could to retrieve attachment file: ' .
+                                    $path .
+                                    ' from storage service: ' .
+                                    $service);
+                            }
+
+                            $content = $result->getContent();
+                            $content = base64_decode(array_get($content, 'content', ''));
+                            $fileName = basename($path);
+                            $filePath = sys_get_temp_dir() . '/' . $fileName;
+                            file_put_contents($filePath, $content);
+                            $attachment[] = new Attachment($filePath, $fileName);
+                        } else {
+                            throw new ForbiddenException(
+                                'You do not have enough privileges to access file: ' .
+                                $path .
+                                ' in service ' .
+                                $service);
+                        }
+                    }
+                }
+            } catch (\Exception $e) {
+                throw new InternalServerErrorException('Failed to get attachment file from storage service. ' .
+                    $e->getMessage());
+            }
+        }
+
+        return $attachment;
+    }
+
+    /**
+     * @param null|string $path
+     *
+     * @return array|mixed|string
+     * @throws \DreamFactory\Core\Exceptions\BadRequestException
+     */
+    protected function getAttachments($path = null)
+    {
+        return array_merge(
+            $this->getUploadedAttachment($path),
+            $this->getUrlAttachment($path),
+            $this->getServiceAttachment($path)
+        );
+    }
+
+    /**
      * @return array
      * @throws BadRequestException
      * @throws NotFoundException
@@ -90,8 +254,11 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
     protected function handlePOST()
     {
         $data = $this->getPayloadData();
-        $templateName = $this->request->getParameter('template', $this->request->getPayloadData('template'));
-        $templateId = $this->request->getParameter('template_id', $this->request->getPayloadData('template_id'));
+        if (empty($data)) {
+            $data = $this->request->input();
+        }
+        $templateName = $this->request->input('template');
+        $templateId = $this->request->input('template_id');
         $templateData = [];
 
         if (!empty($templateName)) {
@@ -118,6 +285,14 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
 
         $data = array_merge((array)array_get($templateData, 'defaults', []), $data);
         $data = array_merge($this->parameters, $templateData, $data);
+
+        // Get attachments from the http request.
+        $attachments = $this->getAttachments();
+        if (!empty($attachments)) {
+            $data['attachment'] = $attachments;
+        } elseif (isset($data['attachment'])) { // Attachment is set in template
+            $data['attachment'] = $this->getAttachments($data['attachment']);
+        }
 
         $text = array_get($data, 'body_text');
         $html = array_get($data, 'body_html');
@@ -164,6 +339,7 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
                 $fromEmail = array_get($data, 'from_email');
                 $replyName = array_get($data, 'reply_to_name');
                 $replyEmail = array_get($data, 'reply_to_email');
+                $attachment = array_get($data, 'attachment');
 
                 if (empty($fromEmail)) {
                     $fromEmail = config('mail.from.address');
@@ -199,6 +375,35 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
                 if (!empty($subject)) {
                     Session::replaceLookups($subject);
                     $m->subject(static::applyDataToView($subject, $data));
+                }
+
+                if (!empty($attachment)) {
+                    if (!is_array($attachment)) {
+                        $attachment = [$attachment];
+                    }
+                    foreach ($attachment as $att) {
+                        if ($att instanceof Attachment) {
+                            $m->attachData($att->getContent(), $att->getName());
+                            $att->unlink();
+                        }
+                    }
+                }
+
+                if (!empty($attachment) && is_array($attachment)) {
+                    if (isset($attachment['name'])) {
+                        $attachments[] = $attachment;
+                    } else {
+                        $attachments = $attachment;
+                    }
+                    foreach ($attachments as $att) {
+                        $fileName = array_get($att, 'name');
+                        $filePath = array_get($att, 'tmp_name');
+                        if (is_file($filePath)) {
+                            $m->attachData(file_get_contents($filePath), $fileName);
+                            // Delete from local /tmp directory.
+                            @unlink($filePath);
+                        }
+                    }
                 }
 
                 if (!empty($bcc)) {
@@ -270,6 +475,13 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
                             'description' => 'Optional template id to base email on.',
                             'schema'      => ['type' => 'integer', 'format' => 'int32'],
                             'in'          => 'query',
+                        ],
+                        [
+                            'name'        => 'attachment',
+                            'description' => 'Import file(s) from URL for attachment. This is also available in form-data post and in json payload data.',
+                            'type'        => 'string',
+                            'in'          => 'query',
+                            'required'    => false,
                         ],
                     ],
                     'requestBody' => [
@@ -393,6 +605,23 @@ abstract class BaseService extends BaseRestService implements EmailServiceInterf
                     'reply_to_email' => [
                         'type'        => 'string',
                         'description' => 'Optional reply to email.',
+                    ],
+                    'attachment'     => [
+                        'type'        => 'array',
+                        'description' => 'File(s) to import from storage service or URL for attachment',
+                        'items'       => [
+                            'type'       => 'object',
+                            'properties' => [
+                                'service' => [
+                                    'type'        => 'string',
+                                    'description' => 'Name of the storage service to use.'
+                                ],
+                                'path'    => [
+                                    'type'        => 'string',
+                                    'description' => 'File path relative to the service.'
+                                ]
+                            ]
+                        ]
                     ],
                 ],
             ],