+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- DF-1191 Added attachment support for email services and email templates
 
 ## [0.6.0] - 2017-08-17
 ### Changed